bitCollector_main
Housekeeping
Renamed main_logger to root_logger for consistency
Added logging to the framework cleanup

    Test1

Housekeeping
Removed the if statement from the end of the module since it is not meant to be run independently of the framework. You can use CLA to initialize a module, but it is preferrable for testing to use the framework.
Renamed main_logger to root_logger
Added the module cleanup method.
Showed an example of a temp file which needed to be cleaned up.